Fix version file syntax error (#37)
Hey @JonnyOThan, the latest TUFX is failing to index in CKAN because
there's an extra set of single quotes inside the value of
`KSP_VERSION_MAX`.

`"'1.12'"`

Looks like this happened in ea4475692255804ba72dee4995c2c952c44077f7,
which suggests KSPBuildTools may be responsible. Probably worth fixing
that if true.